refactor: Split bunch of accounts services into separate files
