+++ src/models/Accounts.model.ts
@@ -168,12 +168,7 @@ export default class Accounts extends Model {
   }
 }
 
-export interface GetAccountsPayload {
-  userId: AccountsAttributes['userId'];
-  type?: AccountsAttributes['type'];
-}
-
-export const getAccounts = async (payload: GetAccountsPayload) => {
+export const getAccounts = async (payload: { userId: number; type?: ACCOUNT_TYPES }) => {
   const { userId, type } = payload;
   const where: {
     userId: AccountsAttributes['userId'];
@@ -204,14 +199,13 @@ export const getAccountById = async ({
   return account;
 };
 
-export interface GetAccountsByExternalIdsPayload {
-  userId: AccountsAttributes['userId'];
-  externalIds: string[];
-}
 export const getAccountsByExternalIds = async ({
   userId,
   externalIds,
-}: GetAccountsByExternalIdsPayload) => {
+}: {
+  userId: AccountsAttributes['userId'];
+  externalIds: string[];
+}) => {
   const account = await Accounts.findAll({
     where: {
       userId,
