+++ src/services/accounts/index.ts
@@ -1,0 +1,2 @@
+export * from './get-accounts';
+export * from './create-account';
